+++ src/TrainingArchitect/Program.cs
@@ -260,6 +260,7 @@ builder.Services.AddSingleton(sp =>
 builder.Services.AddSingleton<IArticleRepository, ArticleRepository>();
 builder.Services.AddSingleton<IProjectRepository, ProjectRepository>();
 builder.Services.AddSingleton<IAboutRepository, AboutRepository>();
+builder.Services.AddSingleton<IUsageCounterRepository, UsageCounterRepository>();
 
 builder.Services.AddScoped<IOwnerService, OwnerService>();
 builder.Services.AddScoped<IAthleteDataService, McpAthleteDataService>();
@@ -471,6 +472,7 @@ app.MapGet("/api/user", (HttpContext context) =>
 app.MapAboutEndpoints();
 app.MapProjectEndpoints();
 app.MapArticleEndpoints();
+app.MapAdminEndpoints();
 app.MapImageEndpoints();
 app.MapAthleteDataEndpoints();
 app.MapCoachEndpoints();